test: LatestAuctionResults
